Move component MonitoringItemsOverviewCard from 'components/cards' to 'components/index/CardsMonitoring'
