+++ components/index/CardsMonitoring/MonitoringItemsOverviewCard.vue
